pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-data-perimeter"
version = "0.1.0"
description = "KMS-centric AWS data perimeter policy generator"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
dp-generate = "generator.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
